Create from abstraction ResponsiveTabsControl
