+++ src/components/responsive-tabs-control/editor.scss
@@ -1,0 +1,19 @@
+.maxi-responsive-tabs-control {
+	.maxi-tabs-content:not(.maxi-tabs-content--disable-padding) {
+		padding-right: 0 !important;
+		padding-bottom: 0 !important;
+		padding-left: 0 !important;
+	}
+	.maxi-tabs-control {
+		margin-left: -12px;
+		margin-right: -12px;
+
+		&__button {
+			border-top: 1px solid #dddfe1;
+			border-right: 1px solid #dddfe1;
+			&:last-of-type {
+				border-right: none;
+			}
+		}
+	}
+}
