--- src/components/typography-control/editor.scss
@@ -1,21 +1,0 @@
-.maxi-typography-control {
-	.maxi-typography-control__text-options-tabs {
-		.maxi-tabs-content {
-			padding-right: 0 !important;
-			padding-bottom: 0 !important;
-			padding-left: 0 !important;
-		}
-		.maxi-tabs-control {
-			margin-left: -12px;
-			margin-right: -12px;
-
-			&__button {
-				border-top: 1px solid #dddfe1;
-				border-right: 1px solid #dddfe1;
-				&:last-of-type {
-					border-right: none;
-				}
-			}
-		}
-	}
-}
